processing-service/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
google-cloud-storage==2.18.0
Pillow==11.0.0
